v 0.0.1

+++ src/queue.c
@@ -1,5 +1,6 @@
-#include "mscompress.h"
+#include <stdio.h>
 #include <stdlib.h>
+#include "mscompress.h"
 
 cmp_blk_queue_t*
 alloc_cmp_buff()
@@ -261,7 +262,7 @@ dump_block_len_queue(block_len_queue_t* queue, int fd)
 }
 
 block_len_queue_t*
-read_block_len_queue(void* input_map, int offset, int end)
+read_block_len_queue(void* input_map, long offset, long end)
 {
     if(input_map == NULL)
         error("read_block_len_queue: input_map is NULL");
